Revert other renames
